fix windows compilation error
- Fixed an issue where we would not be able to compile for windows
