+++ src/helpers/os.go
@@ -14,7 +14,6 @@ import (
 	"path/filepath"
 	"runtime"
 	"strings"
-	"syscall"
 	"time"
 
 	"github.com/briandowns/spinner"
@@ -331,34 +330,6 @@ func FileExists(path string) bool {
 	return true
 }
 
-func extractVolumeName(path string) (string, error) {
-	if runtime.GOOS == "windows" {
-		return "", fmt.Errorf("not supported on windows")
-	}
-	if runtime.GOOS == "linux" {
-		return "", fmt.Errorf("not supported on linux")
-	}
-
-	if !strings.HasPrefix(path, "/Volumes/") {
-		volume := filepath.Dir(path)
-		return volume, nil
-	}
-
-	// Clean up the path to remove any redundant separators
-	cleanedPath := filepath.Clean(path)
-
-	// Split the path into parts based on the separator
-	parts := strings.Split(cleanedPath, string(filepath.Separator))
-
-	// Check if the path has enough parts to contain a volume name
-	if len(parts) < 3 || parts[1] != "Volumes" {
-		return "", fmt.Errorf("invalid path or volume not found in path: %s", path)
-	}
-
-	// Return the volume name (third part in the split path)
-	return fmt.Sprintf("/%s/%s", parts[1], parts[2]), nil
-}
-
 // CopyDir recursively copies a directory tree, attempting to preserve permissions.
 // Source directory must exist, destination directory must *not* exist.
 // Symlinks are ignored and skipped.
@@ -394,26 +365,12 @@ func CopyDir(src string, dst string) (err error) {
 
 		// fmt.Printf("Copying folder with macos clone %s, %s\n", src, dst)
 		var cmd Command
-		// if the destination is a mounted volume, we cannot use the clone command
-		srcVolumeName, err := extractVolumeName(src)
-		if err != nil {
-			return err
-		}
-		dstVolumeName, err := extractVolumeName(dst)
+		isSameVolume, err := isSameVolume(src, dst)
 		if err != nil {
 			return err
 		}
 
-		srcInfo, err := os.Stat(srcVolumeName)
-		if err != nil {
-			return err
-		}
-		dstInfo, err := os.Stat(dstVolumeName)
-		if err != nil && !os.IsNotExist(err) {
-			return err
-		}
-
-		if os.IsNotExist(err) || srcInfo.Sys().(*syscall.Stat_t).Dev != dstInfo.Sys().(*syscall.Stat_t).Dev {
+		if !isSameVolume {
 			cmd = Command{
 				Command: "cp",
 				Args:    []string{"-r", src, dst},
